NFDIV-4792: Refactor code
