--- e2e/journeys/manageCases/createCase/solicitor/solicitorCreateCase/aboutApplicant.ts
@@ -1,15 +1,0 @@
-import {Page} from "@playwright/test";
-import {AboutApplicantPage} from "../../../../../pages/manageCase/solicitorCreateCase/aboutApplicant.ts";
-
-interface SolicitorCreateCaseOptions {
-  page: Page;
-}
-
-export class AboutApplicant {
-  public static async aboutApplicant({
-     page,
-  }: SolicitorCreateCaseOptions): Promise<void> {
-
-    await AboutApplicantPage.aboutApplicant(page);
-  }
-}
