--- e2e/journeys/manageCases/createCase/solicitor/solicitorCreateCase/applicantTwoServiceDetails.ts
@@ -1,17 +1,0 @@
-import {Page} from "@playwright/test";
-import {
-  ApplicantTwoServiceDetailsPage
-} from "../../../../../pages/manageCase/solicitorCreateCase/applicantTwoServiceDetails.ts";
-
-interface SolicitorCreateCaseOptions {
-  page: Page;
-}
-
-export class ApplicantTwoServiceDetails {
-  public static async applicantTwoServiceDetails({
-     page,
-  }: SolicitorCreateCaseOptions): Promise<void> {
-
-    await ApplicantTwoServiceDetailsPage.applicantTwoServiceDetails(page);
-  }
-}
