--- e2e/journeys/manageCases/createCase/solicitor/solicitorCreateCase/checkYourAnswersAndSubmit.ts
@@ -1,17 +1,0 @@
-import {Page} from "@playwright/test";
-import {
-  CheckYourAnswersAndSubmitPage
-} from "../../../../../pages/manageCase/solicitorCreateCase/checkYourAnswersAndSubmit.ts";
-
-interface SolicitorCreateCaseOptions {
-  page: Page;
-}
-
-export class CheckYourAnswersAndSubmit {
-  public static async checkYourAnswers({
-     page,
-  }: SolicitorCreateCaseOptions): Promise<void> {
-
-    await CheckYourAnswersAndSubmitPage.checkYourAnswers(page);
-  }
-}
