--- e2e/journeys/manageCases/createCase/solicitor/solicitorCreateCase/howDoYouWantToApplyForDivorce.ts
@@ -1,17 +1,0 @@
-import {Page} from "@playwright/test";
-import {
-  HowDoYouWantToApplyForDivorcePage
-} from "../../../../../pages/manageCase/solicitorCreateCase/howDoYouWantToApplyForDivorce.ts";
-
-interface SolicitorCreateCaseOptions {
-  page: Page;
-}
-
-export class HowDoYouWantToApplyForDivorce {
-  public static async applicationType({
-     page,
-  }: SolicitorCreateCaseOptions): Promise<void> {
-
-    await HowDoYouWantToApplyForDivorcePage.createApplicationTypePage(page);
-  }
-}
