--- e2e/journeys/manageCases/createCase/solicitor/solicitorCreateCase/marriageBrokenDownIrretrievably.ts
@@ -1,17 +1,0 @@
-import {Page} from "@playwright/test";
-import {
-  MarriageBrokenDownIrretrievablyPage
-} from "../../../../../pages/manageCase/solicitorCreateCase/marriageBrokenDownIrretrievably.ts";
-
-interface SolicitorCreateCaseOptions {
-  page: Page;
-}
-
-export class MarriageBrokenDownIrretrievably {
-  public static async marriageBrokenDown({
-                                           page,
-                                         }: SolicitorCreateCaseOptions): Promise<void> {
-
-    await MarriageBrokenDownIrretrievablyPage.marriageBrokenDown(page);
-  }
-}
