--- e2e/journeys/manageCases/createCase/solicitor/solicitorCreateCase/uploadSupportingDocuments.ts
@@ -1,17 +1,0 @@
-import {Page} from "@playwright/test";
-import {
-  UploadSupportingDocumentsPage
-} from "../../../../../pages/manageCase/solicitorCreateCase/uploadSupportingDocuments.ts";
-
-interface SolicitorCreateCaseOptions {
-  page: Page;
-}
-
-export class UploadSupportingDocuments {
-  public static async uploadSupportingDocuments({
-     page,
-  }: SolicitorCreateCaseOptions): Promise<void> {
-
-    await UploadSupportingDocumentsPage.uploadSupportingDocuments(page);
-  }
-}
