+++ e2e/pages/manageCase/solicitorDraftDivorceApplication/aboutApplicant.ts
@@ -53,9 +53,9 @@ export class AboutApplicantPage {
       {elementId: InputFieldElementIds.applicant1Postcode, fieldValue: AboutApplicants.applicant1Postcode},
     ];
 
-    for (const textField of textFields) {
-      await page.fill(textField.elementId, textField.fieldValue);
-    }
+    const textFieldsInput = textFields.map(textField => page.fill(textField.elementId, textField.fieldValue));
+
+    await Promise.all(textFieldsInput);
 
     await page.click(
       `${Selectors.button}:text-is("${CommonContent.findAddressButton}")`,
